Change UserPage related endpoints

+++ ygoons/modules/user/routes.py
@@ -3,7 +3,7 @@ import json
 import os
 
 import flask
-from flask import request, jsonify, make_response, send_from_directory
+from flask import request, jsonify, make_response, abort, send_from_directory
 from flask import current_app as app
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from werkzeug.utils import secure_filename
@@ -34,79 +34,73 @@ def get_user_info(user_id):
         return make_response(jsonify({'user': user}), 200)
 
 
-@blueprint.route('/user/<int:user_id>/info', methods=['PUT'])
-@jwt_required
+@blueprint.route('/user/<string:user_id>/info', methods=['PUT'])
+@aws_amplify_login_required
 def update_user_info(user_id):
     """
     Updates user information stored in tbl_user and tbl_user_info
     :param user_id: user_id of request
     """
-    if user_id == get_jwt_identity()['userId']:
-        updated_data = json.loads(request.data)
-        row_cnt = 0
-        new_username = None
-        if 'username' in updated_data:
-            new_username = updated_data['username']
-            row_cnt += helpers.update_username(user_id, new_username)
-            del updated_data['username']
+    if request.user['user_id'] != user_id:
+        return abort(403)
 
-        if len(updated_data) > 0:
-            row_cnt += helpers.update_user_info(user_id, updated_data)
+    updated_data = json.loads(request.data)
 
-        if row_cnt > 0:
-            flask.g.pymysql_db.commit()
-            if new_username is not None:
-                return make_response(jsonify({'newUsername': new_username}),
-                                     200)
-            else:
-                return make_response(jsonify({'success': True}), 200)
-        else:
-            return make_response(jsonify({'msg': 'Username update failed'}),
-                                 400)
+    if 'username' in updated_data:
+        username_updated = helpers.update_username(user_id, updated_data['username'])
+        del updated_data['username']
     else:
-        return make_response(
-            jsonify({'msg': 'Authentication failed for request'}), 400)
+        username_updated = True # dummy value for check later
 
+    row_cnt = 0
+    if len(updated_data) > 0:
+        row_cnt += helpers.update_user_info(user_id, updated_data)
 
-@blueprint.route('/user/id/username/<string:username>', methods=['GET'])
-def get_user_exists_by_username(username):
-    """
-    Fetches existence of user of the input username and returns user_id if exists
-    :param username: username to check existence of
-    """
-    print(username)
-    query_result = app.config['COGNITO'].list_users(
-        UserPoolId=app.config['COGNITO_USER_POOL_ID'],
-        AttributesToGet=[
-            'username',
-            'preferred_username'
-        ],
-        Limit=1,
-        Filter="preferred_username='{}'".format(username)
-    )
-    print(query_result)
-    if len(query_result['Users']) == 0:
-        return make_response(jsonify({'message': 'Username already uesed'}), 400)
+    if username_updated and row_cnt == len(updated_data):
+        flask.g.pymysql_db.commit()
+        return make_response(jsonify({'success': True}), 200)
     else:
-        user = query_result['Users'][0]
-    # response = client.list_users(
-    #     UserPoolId='string',
-    #     AttributesToGet=[
-    #         'string',
-    #     ],
-    #     Limit=123,
-    #     PaginationToken='string',
-    #     Filter='string'
-    # )
-    # with flask.g.pymysql_db.cursor() as cursor:
-    #     sql = 'SELECT user_id FROM tbl_user WHERE username = %s'
-    #     cursor.execute(sql, username)
-    #     query_result = cursor.fetchone()
-    # if query_result is not None:
-    #     return make_response(jsonify({'userId': query_result[0]}), 200)
-    # else:
-    #     return make_response(jsonify({'userId': None}), 200)
-        return make_response(jsonify({'user': user}), 200)
+        return make_response(jsonify({'msg': 'User Info Update Failed'}), 400)
+
+
+@blueprint.route('/user/username/<string:username>', methods=['GET'])
+@aws_amplify_login_required
+def get_user_by_username(username):
+    with flask.g.pymysql_db.cursor() as cursor:
+        sql = '''
+        SELECT user_id, username, username_set, email, name, location
+        FROM tbl_user NATURAL JOIN tbl_user_info
+        WHERE username = %s
+        '''
+        cursor.execute(sql, username)
+        query_result = cursor.fetchone()
+    if query_result is not None:
+        user_info = {
+            'username_set': query_result[2],
+            'email': query_result[3],
+            'name': query_result[4],
+            'location': query_result[5]
+        }
+        return make_response(jsonify({'userId': query_result[0], 'userInfo': user_info}), 200)
+    else:
+        return make_response(jsonify({'userId': None, 'userInfo': None}), 200)
+
+
+@blueprint.route('/user/username/<string:username>/id', methods=['GET'])
+def get_user_id_by_username(username):
+    """ Fetches existence of user of the input username and returns user_id if exists
+        :param username: username to check existence of
+    """
+    with flask.g.pymysql_db.cursor() as cursor:
+        sql = '''
+        SELECT user_id FROM tbl_user WHERE username = %s
+        '''
+        cursor.execute(sql, username)
+        query_result = cursor.fetchone()
+    if query_result is not None:
+        return make_response(jsonify({'userId': query_result[0]}), 200)
+    else:
+        return make_response(jsonify({'userId': None}), 200)
 
 
 @blueprint.route('/user/id/email/<string:email>', methods=['GET'])
@@ -119,7 +113,6 @@ def get_user_exists_by_email(email):
         sql = 'SELECT user_id FROM tbl_user WHERE email = %s'
         cursor.execute(sql, email)
         query_result = cursor.fetchone()
-    print(query_result)
     if query_result is not None:
         return make_response(jsonify({'userId': query_result[0]}), 200)
     else:
@@ -162,21 +155,69 @@ def get_user_feed():
     return make_response(jsonify({'postIdArr': post_id_list}), 200)
 
 
-# TODO: error-handling for insert? make sure insert is actually executed
-#       disable following yourself
-@blueprint.route('/user/follow/<int:followed_user_id>', methods=['POST'])
-@jwt_required
+@blueprint.route('/user/<string:user_id>/following', methods=['GET'])
+@aws_amplify_login_required
+def get_following(user_id):
+    """
+    Obtain the list of ids of the users that the user of user_id is following
+    :param user_id:
+    """
+    with flask.g.pymysql_db.cursor() as cursor:
+        sql = '''
+        SELECT followed_id, username
+        FROM (SELECT followed_id FROM tbl_follow WHERE follower_id = %s) tbl_user_follow
+        INNER JOIN tbl_user ON followed_id = user_id
+        '''
+        cursor.execute(sql, (user_id, ))
+        query_result = cursor.fetchall()
+
+    following_list = []
+    for row in query_result:
+        following_list.append(row[0])
+
+    return make_response(jsonify({'followingArr': following_list}), 200)
+
+
+@blueprint.route('/user/<string:user_id>/followers', methods=['GET'])
+@aws_amplify_login_required
+def get_followers(user_id):
+    """
+    Obtain the list of ids of the users that follow the user of user_id
+    :param user_id:
+    """
+    with flask.g.pymysql_db.cursor() as cursor:
+        sql = '''
+        SELECT follower_id, username
+        FROM (SELECT follower_id FROM tbl_follow WHERE followed_id = %s) tbl_user_follow
+        INNER JOIN tbl_user ON follower_id = user_id
+        '''
+        cursor.execute(sql, (user_id, ))
+        query_result = cursor.fetchall()
+
+    follower_list = []
+    for row in query_result:
+        follower_list.append(row[0])
+
+    return make_response(jsonify({'followerArr': follower_list}), 200)
+
+
+@blueprint.route('/user/follow/<string:followed_user_id>', methods=['POST'])
+@aws_amplify_login_required
 def user_follow(followed_user_id):
     """
     Follow the user of given user_id
     :param followed_user_id: user_id of the user to follow
     """
-    curr_user = get_jwt_identity()
-    curr_user_id = curr_user['userId']
+    if request.user['user_id'] == followed_user_id:
+        return abort(400)
+
     with flask.g.pymysql_db.cursor() as cursor:
-        sql = 'INSERT INTO tbl_follow (followed_id, follower_id) ' \
-              'VALUES (%s, %s)'
-        row_cnt = cursor.execute(sql, (followed_user_id, curr_user_id))
+        sql = '''
+        INSERT INTO tbl_follow (followed_id, follower_id)
+        VALUES (%s, %s)
+        '''
+        row_cnt = cursor.execute(sql, (followed_user_id, request.user['user_id']))
+
     if row_cnt == 1:
         flask.g.pymysql_db.commit()
         return make_response(jsonify({'success': True}), 200)
@@ -184,64 +225,25 @@ def user_follow(followed_user_id):
         return make_response(jsonify({'msg': "User follow failed"}), 400)
 
 
-# TODO: disabel unfollowing yourself
-@blueprint.route('/user/follow/<int:followed_user_id>', methods=['DELETE'])
-@jwt_required
+@blueprint.route('/user/follow/<string:followed_user_id>', methods=['DELETE'])
+@aws_amplify_login_required
 def user_unfollow(followed_user_id):
     """
     Unfollow the user of given user_id
     :param followed_user_id: user_id of the user to unfollow
-    :return:
     """
-    curr_user = get_jwt_identity()
-    curr_user_id = curr_user['userId']
+    if request.user['user_id'] == followed_user_id:
+        return abort(400)
+
     with flask.g.pymysql_db.cursor() as cursor:
         sql = 'DELETE FROM tbl_follow WHERE followed_id = %s AND follower_id = %s'
-        affected_row_cnt = cursor.execute(sql,
-                                          (followed_user_id, curr_user_id))
-    if affected_row_cnt == 1:
+        row_cnt = cursor.execute(sql, (followed_user_id, request.user['user_id']))
+
+    if row_cnt == 1:
         flask.g.pymysql_db.commit()
         return make_response(jsonify({'success': True}), 200)
     else:
         return make_response(jsonify({'msg': "User unfollow failed"}), 400)
-
-
-@blueprint.route('/user/<int:user_id>/following', methods=['GET'])
-@jwt_required
-def get_following(user_id):
-    """
-    Obtain the list of ids of the users that the user of user_id is following
-    :param user_id:
-    """
-    with flask.g.pymysql_db.cursor() as cursor:
-        sql = 'SELECT followed_id, username FROM ' \
-              '(SELECT followed_id FROM tbl_follow WHERE follower_id = %s) tbl_user_follow ' \
-              'INNER JOIN tbl_user ON followed_id = user_id'
-        cursor.execute(sql, (user_id, ))
-        query_result = cursor.fetchall()
-    following_list = []
-    for row in query_result:
-        following_list.append(row[0])
-    return make_response(jsonify({'followingArr': following_list}), 200)
-
-
-@blueprint.route('/user/<int:user_id>/followers', methods=['GET'])
-@jwt_required
-def get_followers(user_id):
-    """
-    Obtain the list of ids of the users that follow the user of user_id
-    :param user_id:
-    """
-    with flask.g.pymysql_db.cursor() as cursor:
-        sql = 'SELECT follower_id, username FROM ' \
-              '(SELECT follower_id FROM tbl_follow WHERE followed_id = %s) tbl_user_follow ' \
-              'INNER JOIN tbl_user ON follower_id = user_id'
-        cursor.execute(sql, (user_id, ))
-        query_result = cursor.fetchall()
-    follower_list = []
-    for row in query_result:
-        follower_list.append(row[0])
-    return make_response(jsonify({'followerArr': follower_list}), 200)
 
 
 @blueprint.route('/user/ignore/<int:user_id>', methods=['POST'])
@@ -311,144 +313,84 @@ def upload_full_song_history(post_id):
 
 @blueprint.route('/user/<string:user_id>/profile/image', methods=['GET'])
 @aws_amplify_login_required
-def get_user_profile_image(user_id):
-    # try:
-    url = app.config['S3'].generate_presigned_url(
-        'get_object',
-        Params={
-            'Bucket': app.config['S3_BUCKET_IMAGE'],
-            'Key': os.path.join(user_id, 'profileImage.jpeg')
-        },
-        ExpiresIn=100)
-    return make_response(jsonify({'url': url}), 200)
-
-
-@blueprint.route('/user/<int:user_id>/profile/image', methods=['POST'])
-@jwt_required
-def upload_user_profile_image(user_id):
-    curr_user_id = get_jwt_identity()['userId']
-    if user_id != curr_user_id:
-        return make_response(
-            jsonify({
-                'msg': "Authorization failed",
-                'success': False
-            }), 400)
-
-    image = request.files['file']
-
-    image_file_path = os.path.join(str(user_id),
-                                   secure_filename(image.filename))
-    app.config['S3'].upload_fileobj(
-        image,
-        app.config['S3_BUCKET_IMAGE'],
-        image_file_path,
-        ExtraArgs={"ContentType": image.content_type})
-
-    url = app.config['S3'].generate_presigned_url(
-        'get_object',
-        Params={
-            'Bucket': app.config['S3_BUCKET_IMAGE'],
-            'Key': image_file_path
-        },
-        ExpiresIn=100)
-
-    row_cnt = helpers.upload_profile_picture(user_id, image_file_path)
-    if row_cnt == 1:
-        flask.g.pymysql_db.commit()
-
-    return make_response(jsonify({'url': url}), 200)
-
-
-@blueprint.route('/user/<int:user_id>/profile/image', methods=['DELETE'])
-@jwt_required
-def delete_user_profile_image(user_id):
-    curr_user_id = get_jwt_identity()['userId']
-    if user_id != curr_user_id:
-        return make_response(
-            jsonify({
-                'msg': "Authorization failed",
-                'success': False
-            }), 400)
-
-    row_cnt = helpers.delete_profile_picture(user_id)
-    if row_cnt == 1:
-        flask.g.pymysql_db.commit()
-        return make_response(jsonify({'success': True}), 200)
-    else:
-        return make_response(
-            jsonify({'msg': "Error in deleting profile image"}), 400)
-
-
-@blueprint.route('/user/<int:user_id>/header/image', methods=['GET'])
-def get_user_header_image(user_id):
-    query_result = helpers.get_header_picture_path(user_id)
-    if query_result is not None:
+def get_user_profile_image_get_url(user_id):
+    try:
         url = app.config['S3'].generate_presigned_url(
             'get_object',
             Params={
                 'Bucket': app.config['S3_BUCKET_IMAGE'],
-                'Key': query_result
+                'Key': os.path.join(user_id, 'profileImage.jpeg')
+            },
+            ExpiresIn=3600)
+        return make_response(jsonify({'url': url}), 200)
+    except:
+        return abort(400)
+
+
+@blueprint.route('/user/<string:user_id>/profile/image', methods=['POST'])
+@aws_amplify_login_required
+def get_user_profile_image_post_url(user_id):
+    if request.user['user_id'] != user_id:
+        return abort(403)
+
+    try:
+        res = app.config['S3'].generate_presigned_post(app.config['S3_BUCKET_IMAGE'],
+                                                   os.path.join(user_id, 'profileImage.jpeg'))
+        return make_response(jsonify({'res': res}), 200)
+    except:
+        return abort(400)
+
+
+@blueprint.route('/user/<string:user_id>/profile/image', methods=['DELETE'])
+@aws_amplify_login_required
+def delete_user_profile_image(user_id):
+    if request.user['user_id'] != user_id:
+        return abort(403)
+
+    app.config['S3'].delete_object(Bucket=app.config['S3_BUCKET_IMAGE'],
+                                   Key=os.path.join(user_id, 'profileImage.jpeg'))
+    return make_response(jsonify({'success': True}), 200)
+
+
+@blueprint.route('/user/<string:user_id>/header/image', methods=['GET'])
+@aws_amplify_login_required
+def get_user_header_image_get_url(user_id):
+    try:
+        url = app.config['S3'].generate_presigned_url(
+            'get_object',
+            Params={
+                'Bucket': app.config['S3_BUCKET_IMAGE'],
+                'Key': os.path.join(user_id, 'headerImage.jpeg')
             },
             ExpiresIn=100)
         return make_response(jsonify({'url': url}), 200)
-
-    else:
-        return make_response(jsonify({'msg': "user_id is not found"}), 400)
-
-
-@blueprint.route('/user/<int:user_id>/header/image', methods=['POST'])
-@jwt_required
-def upload_user_header_image(user_id):
-    curr_user_id = get_jwt_identity()['userId']
-    if user_id != curr_user_id:
-        return make_response(
-            jsonify({
-                'msg': "Authorization failed",
-                'success': False
-            }), 400)
-
-    image = request.files['file']
-
-    image_file_path = os.path.join(str(user_id),
-                                   secure_filename(image.filename))
-    app.config['S3'].upload_fileobj(
-        image,
-        app.config['S3_BUCKET_IMAGE'],
-        image_file_path,
-        ExtraArgs={"ContentType": image.content_type})
-
-    url = app.config['S3'].generate_presigned_url(
-        'get_object',
-        Params={
-            'Bucket': app.config['S3_BUCKET_IMAGE'],
-            'Key': image_file_path
-        },
-        ExpiresIn=100)
-
-    row_cnt = helpers.upload_header_picture(user_id, image_file_path)
-    if row_cnt == 1:
-        flask.g.pymysql_db.commit()
-    return make_response(jsonify({'url': url}), 200)
+    except:
+        return abort(400, 'error retrieving header image')
 
 
-@blueprint.route('/user/<int:user_id>/header/image', methods=['DELETE'])
-@jwt_required
+@blueprint.route('/user/<string:user_id>/header/image', methods=['POST'])
+@aws_amplify_login_required
+def get_user_header_image_post_url(user_id):
+    if request.user['user_id'] != user_id:
+        return abort(403)
+
+    try:
+        res = app.config['S3'].generate_presigned_post(app.config['S3_BUCKET_IMAGE'],
+                                                       os.path.join(user_id, 'headerImage.jpeg'))
+        return make_response(jsonify({'res': res}), 200)
+    except:
+        return abort(400, 'error uploading header image')
+
+
+@blueprint.route('/user/<string:user_id>/header/image', methods=['DELETE'])
+@aws_amplify_login_required
 def delete_user_header_image(user_id):
-    curr_user_id = get_jwt_identity()['userId']
-    if user_id != curr_user_id:
-        return make_response(
-            jsonify({
-                'msg': "Authorization failed",
-                'success': False
-            }), 400)
+    if request.user['user_id'] != user_id:
+        return abort(403)
 
-    row_cnt = helpers.delete_header_picture(user_id)
-    if row_cnt == 1:
-        flask.g.pymysql_db.commit()
-        return make_response(jsonify({'success': True}), 200)
-    else:
-        return make_response(
-            jsonify({'msg': "Error in deleting profile image"}), 400)
+    app.config['S3'].delete_object(Bucket=app.config['S3_BUCKET_IMAGE'],
+                                   Key=os.path.join(user_id, 'headerImage.jpeg'))
+    return make_response(jsonify({'success': True}), 200)
 
 
 @blueprint.route('/user/<int:user_id>/played', methods=['GET'])
